Move shape func to _algorithm.py
