+++ python/tvm/relay/op/_transform.py
@@ -204,68 +204,3 @@ def take_shape_func(attrs, inputs, out_ndims):
             axis += data_ndim
         assert 0 <= axis < data_ndim
         return [_take_with_axis_shape_func(*inputs, convert(axis), out_ndims[0])]
-
-@script
-def _argwhere_shape_func_2d(condition):
-    out = output_tensor((2, ), "int64")
-    out[0] = int64(0)
-    out[1] = int64(2)
-    for i1 in range(condition.shape[0]):
-        for i2 in range(condition.shape[1]):
-            if condition[i1, i2]:
-                out[0] += int64(1)
-    return out
-
-@script
-def _argwhere_shape_func_3d(condition):
-    out = output_tensor((2, ), "int64")
-    out[0] = int64(0)
-    out[1] = int64(3)
-    for i1 in range(condition.shape[0]):
-        for i2 in range(condition.shape[1]):
-            for i3 in range(condition.shape[2]):
-                if condition[i1, i2, i3]:
-                    out[0] += int64(1)
-    return out
-
-@script
-def _argwhere_shape_func_4d(condition):
-    out = output_tensor((2, ), "int64")
-    out[0] = int64(0)
-    out[1] = int64(4)
-    for i1 in range(condition.shape[0]):
-        for i2 in range(condition.shape[1]):
-            for i3 in range(condition.shape[2]):
-                for i4 in range(condition.shape[3]):
-                    if condition[i1, i2, i3, i4]:
-                        out[0] += int64(1)
-    return out
-
-@script
-def _argwhere_shape_func_5d(condition):
-    out = output_tensor((2, ), "int64")
-    out[0] = int64(0)
-    out[1] = int64(5)
-    for i1 in range(condition.shape[0]):
-        for i2 in range(condition.shape[1]):
-            for i3 in range(condition.shape[2]):
-                for i4 in range(condition.shape[3]):
-                    for i5 in range(condition.shape[4]):
-                        if condition[i1, i2, i3, i4, i5]:
-                            out[0] += int64(1)
-    return out
-
-@_reg.register_shape_func("argwhere", True)
-def argwhere_shape_func(attrs, inputs, out_ndims):
-    """
-    Shape function for argwhere.
-    """
-    if len(inputs[0].shape) == 2:
-        return [_argwhere_shape_func_2d(inputs[0])]
-    elif len(inputs[0].shape) == 3:
-        return [_argwhere_shape_func_3d(inputs[0])]
-    elif len(inputs[0].shape) == 4:
-        return [_argwhere_shape_func_4d(inputs[0])]
-    elif len(inputs[0].shape) == 5:
-        return [_argwhere_shape_func_5d(inputs[0])]
-    return []
